refactor(backend): consolidate tigerbeetle testcontainer code (#1006)
* feat(backend): consolidate tigerbeetle testcontainer code

* refactor: format

* fix: build backend prior to testing to use tb setup in jest setup

* feat: use typescript for jest setup and teardown

* fix: add ts-node to dev deps

* refactor(backend): turn args list into options object

* refactor(backend): make jest teardown js file again b/c of type issues w/ `global`

* feat(backend): also return mapped port

* refactor(backend): only have tb variables in config that are used by index

* refactor(backend): remove duplicate clusterId constant
