+++ packages/backend/jest.config.js
@@ -9,7 +9,7 @@ module.exports = {
   clearMocks: true,
   testTimeout: 15000,
   roots: [`<rootDir>/packages/${packageName}`],
-  globalSetup: `<rootDir>/packages/${packageName}/jest.setup.js`,
+  globalSetup: `<rootDir>/packages/${packageName}/jest.setup.ts`,
   globalTeardown: `<rootDir>/packages/${packageName}/jest.teardown.js`,
   testRegex: `(packages/${packageName}/.*/__tests__/.*|\\.(test|spec))\\.tsx?$`,
   moduleDirectories: [
