+++ packages/backend/package.json
@@ -36,6 +36,7 @@
     "rosie": "^2.1.0",
     "testcontainers": "^9.1.1",
     "tmp": "^0.2.1",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   },
   "dependencies": {
